fix(auth): login check
- consolidated in <AuthenticatedPage>
- also fixes bug where it always redirect to / on refresh
