+++ src/components/layout/authenticated-page.tsx
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+import { Outlet, useNavigate } from "react-router-dom";
+
+import { useUserStore } from "@/store/user/user-store-provider";
+
+const AuthenticatedPage: React.FC = () => {
+  const { user, isLoading } = useUserStore((store) => store);
+
+  const navigate = useNavigate();
+  useEffect(() => {
+    if (!user && !isLoading) {
+      console.log("ran");
+      navigate("/login");
+    }
+  }, [user, navigate, isLoading]);
+
+  return <Outlet />;
+};
+
+export default AuthenticatedPage;
